+++ src/lib/components/widgets/datatable/DataTable.svelte
@@ -10,10 +10,10 @@
     import type {PROPERTY_VARIATION_TABLE} from "../../../types/variations";
 
     import type {IKeybindEvent} from "../../../actions/keybind";
+    import {navigate_up, navigate_down} from "../../../actions/keybind";
 
     import {IS_BROWSER} from "../../../util/environment";
     import {debounce} from "../../../util/functional";
-    import {navigate_up, navigate_down} from "../../../util/keybind";
     import {clamp} from "../../../util/math";
 
     import * as Table from "../../display/table";